odin: Start adder/subtractor chains by connecting the first cin to gnd/vcc
Instead of burning an adder primitive to create a gnd/vcc for
the first cin in the chain, connect the first cin in the chain
to a global gnd/vcc net.

This type of adder/subtractor chains synthesis is turned
off by default and can turned on using the --adder_cin_global argument.